application-insight-data-generator/requirements.txt
argparse
python-dotenv
opencensus-ext-azure